+++ src/js/handlers.js
@@ -9,9 +9,11 @@
 module.exports = {
   'enter': require( 'handlers/enter' ),
   'go-to-initial-page': require( 'handlers/go-to-initial-page' ),
+  'go-to-password-reset': require( 'handlers/go-to-password-reset' ),
   'authorise-ldap': require( 'handlers/authorise-ldap' ),
   'authorise-google': require( 'handlers/authorise-google' ),
   'authorise-facebook': require( 'handlers/authorise-facebook' ),
+  'send-password-reset-email': require( 'handlers/send-password-reset-email' ),
   'send-passwordless-link': require( 'handlers/send-passwordless-link' )
 };