+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2018-12-14
+### Added
+- Password reset is here!
+
+### Changed
+- Connection for current email is now stored on the email field, as `element` isn't consistent.
+- Tiny change to letter spacing.
+
 ## [1.0.17] - 2018-12-05
 ### Fixed
 - CSP error with Google Analytics